[X]reply 삭제시 실제 파일도 삭제 [X]reply 파일탑재후 work 수정하여 파일탑재할 경우 reply로 submit되는 문제 [X]work 수정시 파일수정안하면 삭제되는 문제 [X]work 삭제시 파일 삭제 안됨 [X]work 삭제후 worklist 끄기 [X]댓글 있을 경우 work 삭제 안됨 메세지 [X]파일 삭제후 work 삭제 작동 안됨 [ ]textarea 엔터키 입력 문제
git-svn-id: https://dev.naver.com/svn/moonndroid@50 12b37519-7715-0510-82f6-ba5b270cff09

+++ WeekWork/index_db.php
@@ -94,11 +94,13 @@ if ($_REQUEST['select'] == "complete") {
 
 if ($_REQUEST['select'] == "delWork") {
     $dir = "./files/";
-    $filehash = $_REQUEST['hash'];
 
-    if (!unlink($dir . $filehash)) {
-        echo "file delete error";
-        exit ;
+    if (!empty($_REQUEST['hash'])) {
+        $filehash = $_REQUEST['hash'];
+
+        if (!unlink($dir . $filehash)) {
+            echo "error";
+        }
     }
     mysql_query("DELETE FROM work WHERE work_id=${_REQUEST['work_id']}");
     mysql_query("DELETE FROM files WHERE file_id=${_REQUEST['file_id']}");
